Moved unit tests into package for nicer importing of helper
