+++ energyplus_pet/equipment/wwhp_heating_curve.py
@@ -172,7 +172,7 @@ Rated Source-side Volumetric Flow Rate: {self.rated_source_volume_flow_rate}
     ):
         cb_progress_initialize(3)
         for i in range(3):
-            sleep(1)
+            sleep(0.3)
             cb_progress_increment()
         cb_progress_done('DONE')
         # set some constants
