+++ energyplus_pet/tests/equipment/test_wwhp_heating_curve.py
@@ -1,7 +1,7 @@
 from energyplus_pet.data_manager import CatalogDataManager
 from energyplus_pet.equipment.wwhp_heating_curve import WaterToWaterHeatPumpHeatingCurveFit
 from energyplus_pet.equipment.equip_types import EquipType
-from .equipment_test_helper import EquipmentTestHelper
+from energyplus_pet.tests.equipment.equipment_test_helper import EquipmentTestHelper
 
 
 class TestWWHPHeatingCurve(EquipmentTestHelper):
